SKS based directories
Forgot that an SKS can get updated (if you take the result with the most
recent modified date that is) so using CHK based directories wasn't updating when republishing the same files under the same keyword. So instead of redirecting to the CHK, just leave it as an SSK for everything. We cache SSK->CHK lookups anyway (though that pointless cleaning operation should be eliminated for repeat cases). Unfortunately that cache means the web server has to be restarted before it'll pull any updates off of gnunet. Not sure how best to manage that. This is such a dumb idea.